Moved metadata_templates from underneath samples scope for projects and groups
