+++ test/controllers/concerns/metadata_templates_actions_concern_test.rb
@@ -15,46 +15,46 @@ class MetadataTemplateActionsConcernTest < ActionDispatch::IntegrationTest
   end
 
   test 'group metadata templates index' do
-    get group_samples_metadata_templates_path(@group)
+    get group_metadata_templates_path(@group)
 
     assert_response :success
   end
 
   test 'group metadata templates new' do
-    get new_group_samples_metadata_template_path(@group, format: :turbo_stream)
+    get new_group_metadata_template_path(@group, format: :turbo_stream)
 
     assert_response :success
   end
 
   test 'group metadata templates edit' do
-    get edit_group_samples_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream)
+    get edit_group_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream)
 
     assert_response :success
   end
 
   test 'group metadata templates show' do
-    get group_samples_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream)
+    get group_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream)
 
     assert_response :success
   end
 
   test 'group metadata templates create' do
     metadata_template_params = { metadata_template: { name: 'Newest template', fields: %w[field1 field5] } }
-    post group_samples_metadata_templates_path(@group, format: :turbo_stream), params: metadata_template_params
+    post group_metadata_templates_path(@group, format: :turbo_stream), params: metadata_template_params
 
     assert_response :success
   end
 
   test 'group metadata templates create failed' do
     metadata_template_params = { metadata_template: { fields: %w[field1 field5] } }
-    post group_samples_metadata_templates_path(@group, format: :turbo_stream), params: metadata_template_params
+    post group_metadata_templates_path(@group, format: :turbo_stream), params: metadata_template_params
 
     assert_response :unprocessable_entity
   end
 
   test 'group metadata templates update' do
     metadata_template_params = { metadata_template: { name: 'This is the new template', fields: %w[field6 field10] } }
-    put group_samples_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream),
+    put group_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream),
         params: metadata_template_params
 
     assert_response :success
@@ -62,48 +62,48 @@ class MetadataTemplateActionsConcernTest < ActionDispatch::IntegrationTest
 
   test 'group metadata templates update failed' do
     metadata_template_params = { metadata_template: { name: nil } }
-    put group_samples_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream),
+    put group_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream),
         params: metadata_template_params
 
     assert_response :unprocessable_entity
   end
 
   test 'group metadata templates destroy' do
-    delete group_samples_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream)
+    delete group_metadata_template_path(@group, @group_metadata_template, format: :turbo_stream)
 
     assert_response :redirect
   end
 
   test 'project metadata templates index' do
-    get namespace_project_samples_metadata_templates_path(@project_namespace.parent, @project)
+    get namespace_project_metadata_templates_path(@project_namespace.parent, @project)
 
     assert_response :success
   end
 
   test 'project metadata templates new' do
-    get new_namespace_project_samples_metadata_template_path(@project_namespace.parent, @project, format: :turbo_stream)
+    get new_namespace_project_metadata_template_path(@project_namespace.parent, @project, format: :turbo_stream)
 
     assert_response :success
   end
 
   test 'project metadata templates edit' do
-    get edit_namespace_project_samples_metadata_template_path(@project_namespace.parent,
-                                                              @project, @project_metadata_template,
-                                                              format: :turbo_stream)
+    get edit_namespace_project_metadata_template_path(@project_namespace.parent,
+                                                      @project, @project_metadata_template,
+                                                      format: :turbo_stream)
 
     assert_response :success
   end
 
   test 'project metadata templates show' do
-    get namespace_project_samples_metadata_template_path(@project_namespace.parent,
-                                                         @project, @project_metadata_template, format: :turbo_stream)
+    get namespace_project_metadata_template_path(@project_namespace.parent,
+                                                 @project, @project_metadata_template, format: :turbo_stream)
 
     assert_response :success
   end
 
   test 'project metadata templates create' do
     metadata_template_params = { metadata_template: { name: 'Newest template', fields: %w[field1 field5] } }
-    post namespace_project_samples_metadata_templates_path(
+    post namespace_project_metadata_templates_path(
       @project_namespace.parent,
       @project, format: :turbo_stream
     ), params: metadata_template_params
@@ -113,7 +113,7 @@ class MetadataTemplateActionsConcernTest < ActionDispatch::IntegrationTest
 
   test 'project metadata templates create failed' do
     metadata_template_params = { metadata_template: { name: 'Newest template' } }
-    post namespace_project_samples_metadata_templates_path(
+    post namespace_project_metadata_templates_path(
       @project_namespace.parent,
       @project, format: :turbo_stream
     ), params: metadata_template_params
@@ -123,7 +123,7 @@ class MetadataTemplateActionsConcernTest < ActionDispatch::IntegrationTest
 
   test 'project metadata templates update' do
     metadata_template_params = { metadata_template: { name: 'This is the new template', fields: %w[field6 field10] } }
-    put namespace_project_samples_metadata_template_path(
+    put namespace_project_metadata_template_path(
       @project_namespace.parent,
       @project, @project_metadata_template, format: :turbo_stream
     ), params: metadata_template_params
@@ -133,7 +133,7 @@ class MetadataTemplateActionsConcernTest < ActionDispatch::IntegrationTest
 
   test 'project metadata templates update failed' do
     metadata_template_params = { metadata_template: { name: nil } }
-    put namespace_project_samples_metadata_template_path(
+    put namespace_project_metadata_template_path(
       @project_namespace.parent,
       @project, @project_metadata_template, format: :turbo_stream
     ), params: metadata_template_params
@@ -142,7 +142,7 @@ class MetadataTemplateActionsConcernTest < ActionDispatch::IntegrationTest
   end
 
   test 'project metadata templates destroy' do
-    delete namespace_project_samples_metadata_template_path(
+    delete namespace_project_metadata_template_path(
       @project_namespace.parent,
       @project, @project_metadata_template,
       format: :turbo_stream
